Extract id visualization + copy functionality as a component
